requirements.txt
argparse>=1.4.0
openai==1.26.0
openpyxl==3.1.1
numpy>=1.24.2
pandas>=1.5.3
python-docx==1.1.2
python-dotenv==1.0.0
tqdm>=4.66.1
webvtt-py==0.5.1
